LPS-95165 Move ServiceWrapperRegistryTest to module adapt to arquillian test
